[ios][maps] Lower minimum OS to iOS 17
