+++ packages/expo-maps/ios/AppleMapsViewiOS18.swift
@@ -1,0 +1,116 @@
+import SwiftUI
+import MapKit
+
+@available(iOS 18.0, *)
+struct AppleMapsViewiOS18: View, AppleMapsViewProtocol {
+  @EnvironmentObject var props: AppleMapsViewProps
+  @ObservedObject private var state = AppleMapsViewState()
+
+  func setCameraPosition(config: CameraPosition?) {
+    withAnimation {
+      state.mapCameraPosition = config.map(convertToMapCamera) ?? .userLocation(fallback: state.mapCameraPosition)
+    }
+  }
+
+  var body: some View {
+    let properties = props.properties
+    let uiSettings = props.uiSettings
+
+    // swiftlint:disable:next closure_body_length
+    MapReader { reader in
+      Map(position: $state.mapCameraPosition, selection: $state.selection) {
+        ForEach(props.markers) { marker in
+          Marker(
+            marker.title,
+            systemImage: marker.systemImage,
+            coordinate: marker.clLocationCoordinate2D
+          )
+          .tint(marker.tintColor)
+          .tag(MapSelection(marker.mapItem))
+        }
+
+        ForEach(props.annotations) { annotation in
+          Annotation(
+            annotation.title,
+            coordinate: annotation.clLocationCoordinate2D
+          ) {
+            ZStack {
+              if let icon = annotation.icon {
+                Image(uiImage: icon.ref)
+                  .resizable()
+                  .frame(width: 50, height: 50)
+              } else {
+                RoundedRectangle(cornerRadius: 5)
+                  .fill(annotation.backgroundColor)
+              }
+              Text(annotation.text)
+                .foregroundStyle(annotation.textColor)
+                .padding(5)
+            }
+          }
+        }
+        UserAnnotation()
+      }
+      .onTapGesture(coordinateSpace: .local) { position in
+        if let coordinate = reader.convert(position, from: .local) {
+          props.onMapClick([
+            "latitude": coordinate.latitude,
+            "longitude": coordinate.longitude
+          ])
+        }
+      }
+      .mapControls {
+        if uiSettings.compassEnabled {
+          MapCompass()
+        }
+        if uiSettings.scaleBarEnabled {
+          MapScaleView()
+        }
+        if uiSettings.togglePitchEnabled {
+          MapPitchToggle()
+        }
+        if uiSettings.myLocationButtonEnabled {
+          MapUserLocationButton()
+        }
+      }
+      .onChange(of: props.cameraPosition) { _, newValue in
+        state.mapCameraPosition = convertToMapCamera(position: newValue)
+      }
+      .onChange(of: state.selection) { _, newValue in
+        if let marker = props.markers.first(where: { $0.mapItem == newValue?.value }) {
+          props.onMarkerClick([
+            "title": marker.title,
+            "tintColor": marker.tintColor,
+            "systemImage": marker.systemImage,
+            "coordinates": [
+              "latitude": marker.coordinates.latitude,
+              "longitude": marker.coordinates.longitude
+            ]
+          ])
+        }
+      }
+      .onMapCameraChange(frequency: .onEnd) { context in
+        let cameraPosition = context.region.center
+        let longitudeDelta = context.region.span.longitudeDelta
+        let zoomLevel = log2(360 / longitudeDelta)
+
+        props.onCameraMove([
+          "coordinates": [
+            "latitude": cameraPosition.latitude,
+            "longitude": cameraPosition.longitude
+          ],
+          "zoom": zoomLevel,
+          "tilt": context.camera.pitch,
+          "bearing": context.camera.heading
+        ])
+      }
+      .mapFeatureSelectionAccessory(props.properties.selectionEnabled ? .automatic : nil)
+      .mapStyle(properties.mapType.toMapStyle(
+        showsTraffic: properties.isTrafficEnabled
+      ))
+      .onAppear {
+        state.mapCameraPosition = convertToMapCamera(position: props.cameraPosition)
+      }
+    }
+  }
+}
